feat(db): phase 1 green — schema, RLS and fixtures verified on bugsha-dev
Applied to a live Supabase project and run end to end:
  15/15 RLS tests pass (38 pgTAP assertions)
  24/24 ledger assertions pass, including L24 global balance

Migrations split by section as docs/03-schema.sql asks. Five further
defects found only by executing them, all logged in docs/DECISIONS.md:
  D1 revised — the coalesce unique index also fails (enum cast is STABLE,
     not IMMUTABLE); replaced with NULLS NOT DISTINCT, which states the
     spec's intent directly
  D10 — the ledger trigger functions carry no search_path and use
     unqualified table names, so every one raises 42P01 the moment it
     fires inside the security-definer functions 01-architecture.md
     mandates. Found by the first ledger post failing.

Also corrected: pg_cron creates its own schema, div() returns numeric,
and RLS test 4 must invite an existing partner_user or it passes on a
foreign-key violation rather than BG100.

+++ supabase/tests/02_ledger.sql
@@ -104,7 +104,7 @@ select is(
   (select e.amount_minor from financial_entry e
    join "order" o on o.order_id = e.order_id
    where o.code = 'D6N-3P' and e.account = 'commission_revenue'),
-  (select div(c.collected_minor * o.commission_bp + 5000, 10000)
+  (select div(c.collected_minor * o.commission_bp + 5000, 10000)::bigint
    from "order" o join cash_collection c on c.order_id = o.order_id
    where o.code = 'D6N-3P'),
   'L6c: commission on a short collection is computed on collected_minor');
@@ -136,7 +136,7 @@ select is(
   'L11: a platform-funded goodwill credit never debits partner_payable');
 
 -- ─── L18. Rounding ──────────────────────────────────────────────────────────
-select is(div(1750::bigint * 2200 + 5000, 10000), 385::bigint,
+select is(div(1750::bigint * 2200 + 5000, 10000)::bigint, 385::bigint,
   'L18: 22% of 1750 is 385, round-half-up at the minor unit');
 
 -- ─── L20/L21. Payout behaviour ──────────────────────────────────────────────
